+++ packages/appcd-config/test/test-config.js
@@ -1,7 +1,8 @@
-import Config, { load } from '../dist/index';
+import Config from '../dist/config';
 import fs from 'fs-extra';
 import path from 'path';
 import tmp from 'tmp';
+
 import { real } from 'appcd-path';
 
 const _tmpDir = tmp.dirSync({
@@ -21,117 +22,6 @@ function makeTempDir() {
 	return dir;
 }
 
-describe('load()', () => {
-	/**
-	 * Creates a `Config` instance and loads the specified configuration and
-	 * environment specific configuration files.
-	 *
-	 * @param {Object} [opts] - Various options.
-	 * @param {Object} [opts.config] - An object with various config settings. The
-	 * config object will be initialized with these values, however if any user-
-	 * defined or environment specific config files are loaded, then this object
-	 * will be re-merged since it always takes precedence.
-	 * @param {String} [opts.configFile] - Path to a config file to load. It may be
-	 * a JavaScript or JSON file.
-	 * @returns {Config}
-	export function load({ config, configFile, defaultConfigFile } = {}) {
-	*/
-	it('should load without any options', () => {
-		const config = load();
-		expect(config).to.be.instanceof(Config);
-		expect(config.toString(0)).to.equal('{}');
-
-		const config2 = load({});
-		expect(config2).to.be.instanceof(Config);
-		expect(config2.toString(0)).to.equal('{}');
-	});
-
-	it('should fail if config option is invalid', () => {
-		expect(() => {
-			load({ config: 'foo' });
-		}).to.throw(TypeError, 'Expected config to be an object');
-
-		expect(() => {
-			load({ config: [] });
-		}).to.throw(TypeError, 'Expected config to be an object');
-	});
-
-	it('should fail if default config file option is invalid', () => {
-		expect(() => {
-			load({ defaultConfigFile: 123 });
-		}).to.throw(TypeError, 'Expected config file to be a string');
-
-		expect(() => {
-			load({ defaultConfigFile: path.join(__dirname, 'noextension') });
-		}).to.throw(Error, 'Config file must be a JavaScript or JSON file');
-
-		const doesnotexist = path.join(__dirname, 'doesnotexist.js');
-		expect(function () {
-			load({ defaultConfigFile: doesnotexist });
-		}).to.throw(Error, `Config file not found: ${doesnotexist}`);
-	});
-
-	it('should fail if config file option is invalid', () => {
-		expect(() => {
-			load({ configFile: 123 });
-		}).to.throw(TypeError, 'Expected config file to be a string');
-
-		expect(() => {
-			load({ configFile: path.join(__dirname, 'noextension') });
-		}).to.throw(Error, 'Config file must be a JavaScript or JSON file');
-
-		const doesnotexist = path.join(__dirname, 'doesnotexist.js');
-		expect(function () {
-			load({ configFile: doesnotexist });
-		}).to.throw(Error, `Config file not found: ${doesnotexist}`);
-	});
-
-	it('should load a config and default config file', () => {
-		const config = load({
-			configFile: path.join(__dirname, 'fixtures', 'good-load.js'),
-			defaultConfigFile: path.join(__dirname, 'fixtures', 'good-default-load.js')
-		});
-		expect(config.toString(0)).to.equal('{"food":["pizza","tacos"],"name":"foo","age":30}');
-	});
-
-	it('should load environment specific config file', () =>  {
-		const config = load({
-			configFile: path.join(__dirname, 'fixtures', 'good-load-foo.js')
-		});
-		expect(config.toString(0)).to.equal('{"environment":{"name":"foo"},"name":"baz"}');
-	});
-
-	it('should merge custom config with config file', () => {
-		const config = load({
-			config: {
-				age: 29,
-				food: 'nachos'
-			},
-			configFile: path.join(__dirname, 'fixtures', 'good-load.js')
-		});
-		expect(config.toString(0)).to.equal('{"food":"nachos","name":"foo","age":29}');
-	});
-
-	it('should not error when initializing readonly settings from config object while loading config file', () => {
-		let config;
-
-		expect(() => {
-			config = load({
-				config: {
-					title: 'baz'
-				},
-				configFile: path.join(__dirname, 'fixtures', 'good-readonly.js')
-			});
-		}).to.not.throw(Error);
-
-		expect(config.get('title')).to.equal('baz');
-
-		expect(() => {
-			config.merge({ title: 'wiz' });
-		}).to.throw(Error, 'Config option "title" is read-only');
-	});
-});
-
 describe('Config', () => {
 	after(() => {
 		fs.removeSync(tmpDir);
@@ -141,7 +31,44 @@ describe('Config', () => {
 		it('should load blank config', () => {
 			const config = new Config();
 			expect(config).to.be.instanceof(Config);
-			expect(config.toString()).to.equal('{}');
+			expect(config.toString()).to.equal([
+				'{',
+				'  "[base]": {},',
+				'  "[user]": {},',
+				'  "[runtime]": {}',
+				'}'
+			].join('\n'));
+		});
+
+		it('should throw exception for bad base config file', () => {
+			expect(function () {
+				new Config({ baseConfigFile: 123 });
+			}).to.throw(TypeError, 'Expected config file to be a string');
+
+			expect(function () {
+				new Config({ baseConfigFile: function () {} });
+			}).to.throw(TypeError, 'Expected config file to be a string');
+
+			expect(function () {
+				new Config({ baseConfigFile: path.join(__dirname, 'noextension') });
+			}).to.throw(Error, 'Config file must be a JavaScript or JSON file');
+
+			expect(function () {
+				new Config({ baseConfigFile: path.join(__dirname, 'fixtures', 'bad.json') });
+			}).to.throw(Error, /^Failed to load config file: /);
+
+			expect(function () {
+				new Config({ baseConfigFile: path.join(__dirname, 'fixtures', 'bad-syntax.js') });
+			}).to.throw(Error, /^Failed to load config file: /);
+
+			// NOTE: this test will pass if Node ever supports ECMAScript 2015 modules
+			expect(function () {
+				new Config({ baseConfigFile: path.join(__dirname, 'fixtures', 'bad-es-export.js') });
+			}).to.throw(Error, /^Failed to load config file: /);
+
+			expect(function () {
+				new Config({ baseConfigFile: path.join(__dirname, 'fixtures', 'bad-export.js') });
+			}).to.throw(Error, 'Expected config file to export an object');
 		});
 
 		it('should throw exception for bad config file', () => {
@@ -156,11 +83,6 @@ describe('Config', () => {
 			expect(function () {
 				new Config({ configFile: path.join(__dirname, 'noextension') });
 			}).to.throw(Error, 'Config file must be a JavaScript or JSON file');
-
-			const doesnotexist = path.join(__dirname, 'doesnotexist.js');
-			expect(function () {
-				new Config({ configFile: doesnotexist });
-			}).to.throw(Error, `Config file not found: ${doesnotexist}`);
 
 			expect(function () {
 				new Config({ configFile: path.join(__dirname, 'fixtures', 'bad.json') });
@@ -180,6 +102,16 @@ describe('Config', () => {
 			}).to.throw(Error, 'Expected config file to export an object');
 		});
 
+		it('should throw exception for bad base config', () => {
+			expect(() => {
+				new Config({ baseConfig: 'foo' });
+			}).to.throw(TypeError, 'Expected base config to be an object');
+
+			expect(() => {
+				new Config({ baseConfig: [] });
+			}).to.throw(TypeError, 'Expected base config to be an object');
+		});
+
 		it('should throw exception for bad config', () => {
 			expect(() => {
 				new Config({ config: 'foo' });
@@ -192,12 +124,12 @@ describe('Config', () => {
 
 		it('should load json config file', () => {
 			const config = new Config({ configFile: path.join(__dirname, 'fixtures', 'good.json') });
-			expect(config.toString(0)).to.equal('{"name":"foo","age":21,"food":["pizza","tacos"]}');
+			expect(config.toString(0)).to.equal('{"[base]":{},"[user]":{},"[runtime]":{"name":"foo","age":21,"food":["pizza","tacos"]}}');
 		});
 
 		it('should load js config file', () => {
 			const config = new Config({ configFile: path.join(__dirname, 'fixtures', 'good.js') });
-			expect(config.toString(0)).to.equal('{"name":"foo","age":21,"food":["pizza","tacos"]}');
+			expect(config.toString(0)).to.equal('{"[base]":{},"[user]":{},"[runtime]":{"name":"foo","age":21,"food":["pizza","tacos"]}}');
 		});
 
 		it('should not error when js config does not export anything', () => {
@@ -228,6 +160,139 @@ describe('Config', () => {
 
 		it('should not error loading an empty config file', () => {
 			new Config({ configFile: path.join(__dirname, 'fixtures', 'empty.js') });
+		});
+	});
+
+	describe('has()', () => {
+		it('should error if key is not a valid string', () => {
+			const config = new Config();
+
+			expect(() => {
+				config.has();
+			}).to.throw(TypeError, 'Expected key to be a string');
+
+			expect(() => {
+				config.has(123);
+			}).to.throw(TypeError, 'Expected key to be a string');
+		});
+
+		it('should check if a value is set', () => {
+			const config = new Config({
+				baseConfig: {
+					a: 'A'
+				},
+				config: {
+					b: {
+						c: 'C'
+					},
+					e: true
+				}
+			});
+
+			expect(config.has('a')).to.be.true;
+			expect(config.has('b')).to.be.true;
+			expect(config.has('b.c')).to.be.true;
+			expect(config.has('b.d')).to.be.false;
+			expect(config.has('e.f.g')).to.be.false;
+			expect(config.has('i')).to.be.false;
+			config.set('i', 'I');
+			expect(config.has('i')).to.be.true;
+		});
+	});
+
+	describe('load()', () => {
+		it('should error if file is invalid', () => {
+			const config = new Config();
+
+			expect(() => {
+				config.load();
+			}).to.throw(TypeError, 'Expected config file to be a string');
+
+			expect(() => {
+				config.load(123);
+			}).to.throw(TypeError, 'Expected config file to be a string');
+		});
+
+		it('should return if file to load is falsey and skipping non-existent', () => {
+			const config = new Config();
+			expect(config.load(null, { skipIfNotExists: true })).to.equal(config);
+		});
+
+		it('should error if file does not have .js or .json extension', () => {
+			const config = new Config();
+			expect(() => {
+				config.load(path.join(__dirname, 'does_not_exist'));
+			}).to.throw(Error, 'Config file must be a JavaScript or JSON file');
+		});
+
+		it('should error if file does not exist', () => {
+			const config = new Config();
+			const file = path.join(__dirname, 'does_not_exist.json');
+			expect(() => {
+				config.load(file);
+			}).to.throw(Error, `Config file not found: ${file}`);
+		});
+
+		it('should not error if file does not exist and skipping non-existent', () => {
+			const config = new Config();
+			const file = path.join(__dirname, 'does_not_exist.json');
+			expect(config.load(file, { skipIfNotExists: true })).to.equal(config);
+		});
+
+		it('should unload namespace if loading again', () => {
+			const config = new Config();
+			config.load(path.join(__dirname, 'fixtures', 'good.json'), { namespace: 'good' });
+			expect(config.toString(0)).to.equal('{"[base]":{},"good":{"good":{"name":"foo","age":21,"food":["pizza","tacos"]}},"[user]":{},"[runtime]":{}}');
+
+			config.load(path.join(__dirname, 'fixtures', 'good2.json'), { namespace: 'good' });
+			expect(config.toString(0)).to.equal('{"[base]":{},"good":{"good":{"name":"bar","age":32,"food":["nachos"]}},"[user]":{},"[runtime]":{}}');
+		});
+	});
+
+	describe('unload()', () => {
+		it('should error if unloading base namespace', () => {
+			const config = new Config();
+			expect(() => {
+				config.unload(Config.Base);
+			}).to.throw(Error, 'Not allowed to unload base namespace');
+		});
+
+		it('should error if unloading user namespace', () => {
+			const config = new Config();
+			expect(() => {
+				config.unload(Config.User);
+			}).to.throw(Error, 'Not allowed to unload user namespace');
+		});
+
+		it('should error if unloading runtime namespace', () => {
+			const config = new Config();
+			expect(() => {
+				config.unload(Config.Runtime);
+			}).to.throw(Error, 'Not allowed to unload runtime namespace');
+		});
+
+		it('should error if namespace is invalid', () => {
+			const config = new Config();
+
+			expect(() => {
+				config.unload();
+			}).to.throw(TypeError, 'Expected namespace to be a string');
+
+			expect(() => {
+				config.unload(123);
+			}).to.throw(TypeError, 'Expected namespace to be a string');
+		});
+
+		it('should return false if namespace is not found', () => {
+			const config = new Config();
+			expect(config.unload('foo')).to.equal(false);
+		});
+
+		it('should unload a namespace', () => {
+			const config = new Config();
+			config.load(path.join(__dirname, 'fixtures', 'good.json'), { namespace: 'good' });
+			expect(config.toString(0)).to.equal('{"[base]":{},"good":{"good":{"name":"foo","age":21,"food":["pizza","tacos"]}},"[user]":{},"[runtime]":{}}');
+			expect(config.unload('good')).to.equal(true);
 		});
 	});
 
@@ -292,6 +357,14 @@ describe('Config', () => {
 			expect(config.get('foo.bar')).to.deep.equal([ 'baz', 'wiz' ]);
 		});
 
+		it('should not push a duplicate', () => {
+			const config = new Config({ config: { foo: [ 'bar' ] } });
+			config.push('foo', 'bar');
+			expect(config.get('foo')).to.deep.equal([ 'bar' ]);
+			config.push('foo', [ 'bar', 'baz' ]);
+			expect(config.get('foo')).to.deep.equal([ 'bar', 'baz' ]);
+		});
+
 		it('should push overwrite an existing falsey value', () => {
 			const config = new Config({ config: { foo: { bar: null } } });
 			config.push('foo.bar', 'baz');
@@ -325,6 +398,21 @@ describe('Config', () => {
 			expect(() => {
 				config.push('arrays.multi', true);
 			}).to.throw(Error, 'Invalid "arrays.multi" value "true"');
+		});
+
+		it('should push to user-defined value loaded from file', () => {
+			const config = new Config();
+			config.load(path.join(__dirname, 'fixtures', 'good2.json'));
+			config.push('name', 'baz');
+			expect(config.get('name')).to.deep.equal([ 'bar', 'baz' ]);
+			expect(config.toString(0)).to.equal('{"[base]":{},"[user]":{"age":32,"food":["nachos"],"name":["bar","baz"]},"[runtime]":{"age":32,"food":["nachos"],"name":["bar","baz"]}}');
+		});
+
+		it('should push to user-defined array loaded from file', () => {
+			const config = new Config();
+			config.load(path.join(__dirname, 'fixtures', 'good2.json'));
+			config.push('food', 'breadsticks');
+			expect(config.get('food')).to.deep.equal([ 'nachos', 'breadsticks' ]);
 		});
 	});
 
@@ -364,6 +452,16 @@ describe('Config', () => {
 			const config = new Config({ config: { foo: { bar: [ 'baz' ] } } });
 			config.unshift('foo.bar', 'wiz');
 			expect(config.get('foo.bar')).to.deep.equal([ 'wiz', 'baz' ]);
+		});
+
+		it('should not unshift a duplicate', () => {
+			const config = new Config({ config: { foo: [ 'bar' ] } });
+			config.unshift('foo', 'bar');
+			expect(config.get('foo')).to.deep.equal([ 'bar' ]);
+			config.unshift('foo', [ 'bar', 'baz' ]);
+			expect(config.get('foo')).to.deep.equal([ 'baz', 'bar' ]);
+			config.unshift('foo', [ 'bar', 'foo', 'baz', 'wiz' ]);
+			expect(config.get('foo')).to.deep.equal([ 'foo', 'wiz', 'baz', 'bar' ]);
 		});
 
 		it('should support adding multiple values', () => {
@@ -462,7 +560,7 @@ describe('Config', () => {
 			const config = new Config();
 			config.set('foo', 'bar');
 			expect(config.get('foo')).to.equal('bar');
-			expect(config.toString(0)).to.equal('{"foo":"bar"}');
+			expect(config.toString(0)).to.equal('{"[base]":{},"[user]":{"foo":"bar"},"[runtime]":{"foo":"bar"}}');
 		});
 
 		it('should set/get a deep property', () => {
@@ -470,7 +568,7 @@ describe('Config', () => {
 			config.set('hello.world', 'foo');
 			expect(config.get('hello')).to.deep.equal({ world: 'foo' });
 			expect(config.get('hello.world')).to.equal('foo');
-			expect(config.toString(0)).to.equal('{"hello":{"world":"foo"}}');
+			expect(config.toString(0)).to.equal('{"[base]":{},"[user]":{"hello":{"world":"foo"}},"[runtime]":{"hello":{"world":"foo"}}}');
 		});
 
 		it('should error if key is not a string', () => {
@@ -546,7 +644,7 @@ describe('Config', () => {
 			config.set('foo', 'bar');
 			config.delete('foo');
 			expect(config.get('foo')).to.be.undefined;
-			expect(config.toString(0)).to.equal('{}');
+			expect(config.toString(0)).to.equal('{"[base]":{},"[user]":{},"[runtime]":{}}');
 		});
 
 		it('should delete deep property', () => {
@@ -554,12 +652,12 @@ describe('Config', () => {
 			config.set('foo.bar', 'baz');
 			config.delete('foo');
 			expect(config.get('foo')).to.be.undefined;
-			expect(config.toString(0)).to.equal('{}');
+			expect(config.toString(0)).to.equal('{"[base]":{},"[user]":{},"[runtime]":{}}');
 
 			config.set('foo.bar', 'baz');
 			config.delete('foo.bar');
 			expect(config.get('foo')).to.be.undefined;
-			expect(config.toString(0)).to.equal('{}');
+			expect(config.toString(0)).to.equal('{"[base]":{},"[user]":{},"[runtime]":{}}');
 		});
 
 		it('should return false if deleting non-existent key', () => {
@@ -614,7 +712,7 @@ describe('Config', () => {
 		it('should return immediately if source is not an object', () => {
 			const config = new Config();
 			expect(config.merge()).to.equal(config);
-			expect(config.toString(0)).to.equal('{}');
+			expect(config.toString(0)).to.equal('{"[base]":{},"[user]":{},"[runtime]":{}}');
 		});
 
 		it('should append arrays', () => {
@@ -624,11 +722,20 @@ describe('Config', () => {
 			expect(config.get('foo')).to.deep.equal([ 'bar', 'baz' ]);
 		});
 
+		it('should not merge duplicate array elements', () => {
+			const config = new Config();
+			config.set('foo', [ 'bar' ]);
+			config.merge({ foo: [ 'baz', 'bar' ] });
+			expect(config.get('foo')).to.deep.equal([ 'bar', 'baz' ]);
+			config.merge({ foo: [ 'wiz', 'bar', 'fiz' ] });
+			expect(config.get('foo')).to.deep.equal([ 'bar', 'baz', 'wiz', 'fiz' ]);
+		});
+
 		it('should mix deep objects', () => {
 			const config = new Config();
 			config.set('foo', { bar: 'baz' });
 			config.merge({ foo: { wiz: 'pow' } });
-			expect(config.toString(0)).to.equal('{"foo":{"bar":"baz","wiz":"pow"}}');
+			expect(config.toString(0)).to.equal('{"[base]":{},"[user]":{"foo":{"bar":"baz","wiz":"pow"}},"[runtime]":{"foo":{"bar":"baz","wiz":"pow"}}}');
 		});
 	});
 
@@ -694,11 +801,12 @@ describe('Config', () => {
 
 			fs.writeFileSync(inFile, JSON.stringify(values, null, 4));
 
-			config.loadUserConfig(inFile);
+			config.load(inFile);
 
 			await config.save(outFile);
 
 			expect(JSON.parse(fs.readFileSync(outFile))).to.deep.equal(values);
+			expect(config.toString(0)).to.equal('{"[base]":{},"[user]":{"alpha":"beta","baz":{"wow":"zip","wiz":"wee"},"arr":[1,2,3]},"[runtime]":{"foo":"bar","alpha":"beta","baz":{"wow":"zip","wiz":"wee"},"arr":[1,2,3]}}');
 		});
 	});
 
@@ -734,7 +842,7 @@ describe('Config', () => {
 	describe('metadata', () => {
 		it('should load js config file with metadata', () => {
 			const config = new Config({ configFile: path.join(__dirname, 'fixtures', 'good-meta.js') });
-			expect(config.toString(0)).to.equal('{"name":"foo","age":21,"id":"123456","arrays":{"simple":["pizza","tacos"],"arrayNums":[1,2,3],"arrayNums2":[4,5,6],"arrayNums3":[7,8,9],"multi":["a",1],"multiUnknown":["a"]},"multi":null,"couldBeNull":null,"couldBeNull2":null,"cantBeNull":0,"obsolete":true,"job":{"title":"coder"},"foo":"bar"}');
+			expect(config.toString(0)).to.equal('{"[base]":{},"[user]":{},"[runtime]":{"name":"foo","age":21,"id":"123456","arrays":{"simple":["pizza","tacos"],"arrayNums":[1,2,3],"arrayNums2":[4,5,6],"arrayNums3":[7,8,9],"multi":["a",1],"multiUnknown":["a"]},"multi":null,"couldBeNull":null,"couldBeNull2":null,"cantBeNull":0,"obsolete":true,"job":{"title":"coder"},"foo":"bar"}}');
 
 			validateMetadata(config.meta.get('name'), 'Everybody has a name.', 'String', false, false);
 			validateMetadata(config.meta.get('age'), 'How old are you?', 'Number', 'Age ain\'t nothing but a number', false);
@@ -756,7 +864,7 @@ describe('Config', () => {
 
 		it('should load simple js config file with metadata', () => {
 			const config = new Config({ configFile: path.join(__dirname, 'fixtures', 'good-meta-simple.js') });
-			expect(config.toString(0)).to.equal('{"name":"foo","age":21,"id":"123456","arrays":{"simple":["pizza","tacos"],"arrayNums":[1,2,3],"arrayNums2":[4,5,6],"arrayNums3":[7,8,9],"multi":["a",1]},"multi":null,"couldBeNull":null,"couldBeNull2":null,"cantBeNull":0,"job":{"title":"coder"},"noDesc":"nada"}');
+			expect(config.toString(0)).to.equal('{"[base]":{},"[user]":{},"[runtime]":{"name":"foo","age":21,"id":"123456","arrays":{"simple":["pizza","tacos"],"arrayNums":[1,2,3],"arrayNums2":[4,5,6],"arrayNums3":[7,8,9],"multi":["a",1]},"multi":null,"couldBeNull":null,"couldBeNull2":null,"cantBeNull":0,"job":{"title":"coder"},"noDesc":"nada"}}');
 
 			validateMetadata(config.meta.get('name'), 'Everybody has a name.', 'String', false, false);
 			validateMetadata(config.meta.get('age'), 'How old are you?', 'Number', 'Age ain\'t nothing but a number', false);